+++ README.md
@@ -1,13 +1,15 @@
 # Rustnance 
 
-## A crate that provides functionality within the space of financial analysis
+## Synopsis
+A crate that provides functionality within the space of financial analysis
 
-This crate is only meant to provide calculations, metrics and so on and should not be seen as any financial or investing advice. As of now the intent is to firstly develop calculation for fundamental analysis but technical analysis related functionality is to be expected in the future. Some of the existing functions you will find consist of: 
+## Explanation
+This crate is meant to provide calculations, metrics and so on and should not be seen as any financial or investing advice. As of now the intent is to firstly develop calculation for fundamental analysis but technical analysis related functionality is to be expected in the future. Some of the existing functions you will find consist of: 
 
 * Calculate intrinsic value
 * Calculate compound interest
 
-## How to use
+## Usage
 
 ## Module structure
